Separate test for search
